requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
frozendict==2.4.4
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
multitasking==0.0.11
openai==1.54.0
packaging==24.1
pandas==2.2.3
peewee==3.17.6
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
waitress==3.0.0
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.54
PyPDF2==3.0.1
openpyxl==3.1.5
pycryptodome==3.20.0
qdrant_client==1.12.1
sentence_transformers==3.2.1
langchain==0.3.7
langchain_huggingface==0.1.2
langchain_openai==0.2.8
langchain-community==0.3.7
tortoise-orm==0.24.2
tortoise-orm[asyncodbc]==0.24.2
uvicorn==0.30.5
quart==0.19.9
quart-cors==0.7.0
hypercorn==0.17.3
hyperframe==6.0.1
uvicorn==0.30.5